moving eventlistener to organism_buttonFilledPictures
