+++ generics/components/organisms/Organism_ButtonFilledPictures.mjs
@@ -2,6 +2,7 @@ import { Component } from "../../../core/Component.mjs";
 import { slot } from "../../../core/helpers.mjs";
 import { Molecule_ImageAndText } from "../molecules/Molecule_ImgAndText.mjs";
 import { Atom_ButtonPositive } from "../atoms/Atom_ButtonPositive.mjs";
+import { Modal_ProjectInfo } from "./Modal_ProjectInfo.mjs";
 
 export function Organism_ButtonFilledPictures ( model )
 {
@@ -14,7 +15,10 @@ export function Organism_ButtonFilledPictures ( model )
             <div id="organism_buttonFilledPicture_btnParent">
                 ${ slot( "btn" ) }
             </div>
+            <div class="organism_buttonfilledPic-cards">
             ${ ( model.cards.map( ( card, index ) => slot( "card" + index ) ) ).join( "" ) }
+            </div>
+            <div id="modal-projectInfo">
         </div>
         `;
     };
@@ -29,5 +33,18 @@ export function Organism_ButtonFilledPictures ( model )
             let card = new Molecule_ImageAndText( cardData );
             this.fillSlot( "card" + index, card.getElement() );
         } );
+
+        this.getElement().querySelectorAll("organism_buttonfilledPic-cards").addEventListener("click", (e) => {
+            console.log('btn-project button pressed')      
+            const modalProjInfo = document.getElementById('modal-projectInfo')
+            
+            modalProjInfo.innerHTML = `
+                ${slot("new-modal")}
+            `
+            this.modal = new Modal_ProjectInfo(model.content)
+
+            this.fillSlot("new-modal", this.modal.getElement());
+        });
+
     };
 }
