+++ generics/components/templates/Template_Projects_View.mjs
@@ -19,11 +19,13 @@ export function Template_Projects_View ( view )
             <div class="organism_btn-filled-pictures">
                 ${slot( "organismButtonFilledPictures")}
             </div>
-            <div id="modal-projectInfo">
+            
         </div>`;
     };
 
     // ${slot("projInfoModal")}
+
+    // <div id="modal-projectInfo"></div>
 
     this.bindScript = function ()
     {
@@ -42,17 +44,17 @@ export function Template_Projects_View ( view )
         // let projInfoModal = new Organism_ProjectInfo(model.organism_projectInfo)
         // this.fillSlot("projInfoModal", projInfoModal.getElement())
 
-        this.getElement().querySelectorAll("div.molecule_image-and-text").addEventListener("click", (e) => {
-            console.log('btn-project button pressed')      
-            const modalProjInfo = document.getElementById('modal-projectInfo')
+        // this.getElement().querySelectorAll("div.molecule_image-and-text").addEventListener("click", (e) => {
+        //     console.log('btn-project button pressed')      
+        //     const modalProjInfo = document.getElementById('modal-projectInfo')
             
-            modalProjInfo.innerHTML = `
-                ${slot("new-modal")}
-            `
-            this.modal = new Modal_ProjectInfo(model.content)
+        //     modalProjInfo.innerHTML = `
+        //         ${slot("new-modal")}
+        //     `
+        //     this.modal = new Modal_ProjectInfo(model.content)
 
-            this.fillSlot("new-modal", this.modal.getElement());
-        });
+        //     this.fillSlot("new-modal", this.modal.getElement());
+        // });
 
 
     };
